// api_producao/src/logsSystem/enum/logs.enum.ts
export enum LogsEnum {
  RATEIO_CONCLUIDO = 'RATEIO_CONCLUIDO',
  RATEIO_CONCLUIDO_FAILED = 'RATEIO_CONCLUIDO_FAILED',
  CRIAR_RATEIO = 'CRIAR_RATEIO',
  CRIAR_RATEIO_FAILED ='CRIAR_RATEIO_FAILED',
  USER_CREATED = 'USER_CREATED',
  USER_CREATION_FAILED = 'USER_CREATION_FAILED',
  UPDATE_PASSWORD = 'UPDATE_PASSWORD',
  UPDATE_ACCESSLEVEL = 'UPDATE_ACCESSLEVEL',
  UPDATE_FAILED = 'UPDATE_FAILED',
  DELETE_USER = 'DELETE_USER',
  DELETE_FAILED = 'DELETE_FAILED',
  LINK_USER = 'LINK_USER',
  LINK_FAILED = 'LINK_FAILED'
}